Fix bugs with crash data and add new config to only send crash data on a crash event
Make crash reporter delete crash data after it's sent and only update count on new crashes
Make crash reporter send a crash event and only send data with that event
At the moment, with the default behavior crash data is still sent to all events until the app is closed.
Reason being that someone could call `gatherTrackData` and that would make us report the crash data only on that method
This would cause the crash data to be deleted and never sent in a track request.